warnings fixed
git-svn-id: file:///home/rhornig/tmp/svn/inet-framework/trunk@1827 539eec35-cc16-0410-b457-356aab9780e1

+++ INET/src/mobility/BonnMotionMobility.cc
@@ -65,7 +65,7 @@ void BonnMotionMobility::setTargetPosition()
 {
     const BonnMotionFile::Line& vec = *vecp;
 
-    if (vecpos+2 >= vec.size())
+    if (vecpos+2 >= (int)vec.size())
     {
         stationary = true;
         return;
